iter

+++ Models/Entities/Solicitud.cs
@@ -11,11 +11,11 @@ namespace BackendSWGAF.Models.Entities
 
         public string nombre { get; set; }
 
-        public float cantidad { get; set; }
+        public double cantidad { get; set; }
 
         public string tipoDePago { get; set; }
 
-        public float total { get; set; }
+        public double total { get; set; }
         public DateTime fecha { get; set; }
         public int aceptado { get; set; }
         public int idUsuario { get; set; }
